python3.10 ready

+++ exostriker/lib/calculator.py
@@ -151,8 +151,8 @@ class Calculator(QWidget):
         mainLayout.addWidget(self.addToMemoryButton, 5, 0)
 
         for i in range(1, Calculator.NumDigitButtons):
-            row = ((9 - i) / 3) + 2
-            column = ((i - 1) % 3) + 1
+            row = int(((9 - i) / 3) + 2)
+            column = int(((i - 1) % 3) + 1)
             mainLayout.addWidget(self.digitButtons[i], row, column)
 
         mainLayout.addWidget(self.digitButtons[0], 5, 1)
